Fix For GoTo CLI
When using the goto cli, it uses the +n/-n migrate call. Because it uses an offset with the diff between two timestamps this can cause all migrations to call up/down respectively with a large offset value causing all up/down scripts to be run. Added a goto method that will determine the direction and then stop when the timestamp exceeds (or equals in the case of an upgrade) the desired migration point.